Feat: Implement millisecond and microsecond precision timers

+++ include/state_machine/state_machine_base.h
@@ -49,8 +49,8 @@ typedef struct {
     bool can_run;                // Flag indicating if the robot can run
     uint8_t lap;                 // Current number of laps completed
     uint8_t laps;                // Number of laps to run
-    uint16_t start_time;         // Start time in seconds
-    uint16_t stop_time;          // Time in seconds for the robot to stop
+    uint32_t start_time;         // Time robot started running
+    uint32_t stop_time;          // Time in seconds for the robot to stop
     bool log_data;               // Flag indicating if data should be logged
 } StateMachine;
